refactor of hms
