+++ libbarto/src/lib.rs
@@ -291,10 +291,10 @@ pub use self::message::shared::update::UpdateKind;
 pub use self::message::shared::uuid::UuidWrapper;
 pub use self::schedule::Realtime;
 pub use self::schedule::dow::DayOfWeek;
-pub use self::schedule::hms::Hour;
 pub use self::schedule::hms::HourMinuteSecond;
-pub use self::schedule::hms::Minute;
-pub use self::schedule::hms::Second;
+pub use self::schedule::hms::hour::Hour;
+pub use self::schedule::hms::minute::Minute;
+pub use self::schedule::hms::second::Second;
 pub use self::schedule::ymd::YearMonthDay;
 pub use self::schedule::ymd::day::Day;
 pub use self::schedule::ymd::month::Month;
